# specs/001-warehouse-inventory-system/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Warehouse Inventory Management API
  description: |
    REST API per la gestione del magazzino officina con registrazione movimenti (IN/OUT/ADJUSTMENT),
    inventario in tempo reale, e funzionalità di export Excel.

    ## Architettura
    - **Backend**: FastAPI con validazione Pydantic
    - **Database**: PostgreSQL 15+ con event-sourcing (movimenti immutabili)
    - **Frontend**: Vercel (Next.js)

    ## Caratteristiche Principali
    - Giacenze calcolate in tempo reale da movimenti
    - Validazione movimenti OUT per prevenire giacenze negative
    - Rettifiche inventario (ADJUSTMENT) con nota obbligatoria
    - Export Excel lato client (ultimi 12 mesi)
    - Supporto per quantità decimali (peso/volume)

    ## Localizzazione
    Tutti i messaggi di errore e le descrizioni sono in italiano.
  version: 1.0.0
  contact:
    name: Warehouse Inventory System
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server (localhost)
  - url: https://inventario-magazzino.onrender.com
    description: Production server (Render)
...
paths:
  /api/items:
    get:
      tags: [Items]
      summary: Lista articoli con filtri
      description: |
        Restituisce l'elenco completo degli articoli in magazzino con giacenze calcolate in tempo reale.
        Supporta ricerca per nome, filtro per categoria, e visualizzazione solo articoli sotto scorta.

        **Requisiti**: FR-001, FR-002, FR-003, FR-004
      operationId: listItems
      parameters:
        - name: search
          in: query
          description: Ricerca parziale nel nome articolo (case-insensitive)
          required: false
          schema:
            type: string
            example: "filtro"
        - name: category
          in: query
          description: Filtra per categoria esatta
          required: false
          schema:
            type: string
            example: "Filtri olio"
        - name: under_stock_only
          in: query
          description: Mostra solo articoli sotto scorta minima
          required: false
          schema:
            type: boolean
            default: false
        - name: sort_by
          in: query
          description: Campo per ordinamento
          required: false
          schema:
            type: string
            enum: [name, category, stock_quantity, is_under_min_stock]
            default: name
        - name: sort_order
          in: query
          description: Direzione ordinamento
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Lista articoli restituita con successo
          headers:
            Access-Control-Allow-Origin:
              description: CORS header per frontend Vercel
              schema:
                type: string
                example: "*"
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ItemWithStock'
                  total:
                    type: integer
                    description: Numero totale di articoli (senza filtri)
                    example: 42
                  filtered:
                    type: integer
                    description: Numero di articoli dopo applicazione filtri
                    example: 12
              examples:
                success:
                  summary: Risposta con articoli
                  value:
                    items:
                      - id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                        name: "Olio motore 5W30"
                        category: "Lubrificanti"
                        unit: "lt"
                        notes: "Olio sintetico long life"
                        min_stock: 5.0
                        unit_cost: 8.50
                        stock_quantity: 17.0
                        stock_value: 144.50
                        is_under_min_stock: false
                        last_movement_at: "2025-11-10T14:30:00Z"
                        created_at: "2025-01-15T10:00:00Z"
                      - id: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                        name: "Filtro olio"
                        category: "Filtri"
                        unit: "pz"
                        notes: null
                        min_stock: 10.0
                        unit_cost: 4.20
                        stock_quantity: 5.0
                        stock_value: 21.00
                        is_under_min_stock: true
                        last_movement_at: "2025-11-08T09:15:00Z"
                        created_at: "2025-01-20T11:30:00Z"
                    total: 42
                    filtered: 2
        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      tags: [Items]
      summary: Crea nuovo articolo
      description: |
        Crea un nuovo articolo in magazzino con giacenza iniziale zero.
        I movimenti successivi aggiorneranno la giacenza.

        **Requisiti**: FR-013, FR-020, FR-021
      operationId: createItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreate'
            examples:
              standard:
                summary: Articolo standard
                value:
                  name: "Filtro abitacolo"
                  category: "Filtri"
                  unit: "pz"
                  notes: "Formato universale"
                  min_stock: 5.0
                  unit_cost: 8.50
              no_category:
                summary: Articolo senza categoria
                value:
                  name: "Bulloni M8"
                  unit: "pz"
                  min_stock: 20.0
                  unit_cost: 0.15
              free_item:
                summary: Articolo con costo zero
                value:
                  name: "Materiale recuperato"
                  category: "Varie"
                  unit: "kg"
                  min_stock: 0.0
                  unit_cost: 0.0
                  notes: "Recupero da dismissioni"
      responses:
        '201':
          description: Articolo creato con successo
          headers:
            Location:
              description: URI del nuovo articolo
              schema:
                type: string
                example: "/api/items/a1b2c3d4-e5f6-7890-abcd-ef1234567890"
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: "*"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemWithStock'
              examples:
                success:
                  summary: Articolo creato
                  value:
                    id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    name: "Filtro abitacolo"
                    category: "Filtri"
                    unit: "pz"
                    notes: "Formato universale"
                    min_stock: 5.0
                    unit_cost: 8.50
                    stock_quantity: 0.0
                    stock_value: 0.0
                    is_under_min_stock: true
                    last_movement_at: null
                    created_at: "2025-11-11T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Conflitto - Nome articolo già esistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicate:
                  summary: Nome duplicato
                  value:
                    detail: "Esiste già un articolo con nome 'Filtro abitacolo'"
                    error_code: "DUPLICATE_ITEM_NAME"
  /api/items/{id}:
    get:
      tags: [Items]
      summary: Dettaglio articolo
      description: Restituisce i dettagli completi di un articolo specifico con giacenza corrente
      operationId: getItem
      parameters:
        - $ref: '#/components/parameters/ItemId'
      responses:
        '200':
          description: Articolo trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemWithStock'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Items]
      summary: Aggiorna articolo
      description: |
        Modifica i dati di un articolo esistente. La giacenza NON può essere modificata
        direttamente (si usa registrazione movimenti).

        **Requisiti**: FR-014
      operationId: updateItem
      parameters:
        - $ref: '#/components/parameters/ItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemUpdate'
            examples:
              update_cost:
                summary: Aggiorna costo unitario
                value:
                  name: "Olio motore 5W30"
                  category: "Lubrificanti"
                  unit: "lt"
                  notes: "Olio sintetico long life - nuovo fornitore"
                  min_stock: 5.0
                  unit_cost: 9.00
              update_threshold:
                summary: Modifica soglia minima
                value:
                  name: "Filtro olio"
                  category: "Filtri"
                  unit: "pz"
                  min_stock: 15.0
                  unit_cost: 4.20
      responses:
        '200':
          description: Articolo aggiornato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemWithStock'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Items]
      summary: Elimina articolo
      description: |
        Elimina un articolo dal magazzino. Permesso solo se:
        - Giacenza corrente = 0
        - Nessun movimento negli ultimi 12 mesi

        **Requisiti**: FR-015, FR-016, FR-017
      operationId: deleteItem
      parameters:
        - $ref: '#/components/parameters/ItemId'
      responses:
        '204':
          description: Articolo eliminato con successo
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflitto - Eliminazione non permessa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                non_zero_stock:
                  summary: Giacenza non zero
                  value:
                    detail: "Impossibile eliminare: giacenza non zero (attuale: 12.5 pz)"
                    error_code: "CANNOT_DELETE_NON_ZERO_STOCK"
                recent_movements:
                  summary: Movimenti recenti
                  value:
                    detail: "Impossibile eliminare: l'articolo ha movimenti recenti (ultimi 12 mesi). Ultimo movimento: 2025-09-15"
                    error_code: "CANNOT_DELETE_HAS_RECENT_MOVEMENTS"
  /api/items/autocomplete/categories:
    get:
      tags: [Items]
      summary: Autocomplete categorie
      description: |
        Restituisce l'elenco delle categorie già utilizzate per suggerimenti autocomplete.

        **Requisiti**: FR-018
      operationId: autocompleteCategories
      parameters:
        - name: q
          in: query
          description: Filtro parziale per la categoria
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Lista categorie
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      type: string
                    example: ["Filtri", "Filtri olio", "Filtri aria", "Lubrificanti", "Freni", "Liquidi"]
...
components:
  schemas:
    ItemCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Nome articolo (univoco)
          example: "Filtro abitacolo"
        category:
          type: string
          maxLength: 100
          nullable: true
          description: Categoria libera (con autocomplete)
          example: "Filtri"
        unit:
          type: string
          maxLength: 20
          default: "pz"
          description: Unità di misura
          example: "pz"
        notes:
          type: string
          nullable: true
          description: Note descrittive
          example: "Formato universale, compatibile maggior parte auto"
        min_stock:
          type: number
          format: decimal
          minimum: 0
          default: 0
          description: Soglia minima per alert sotto scorta
          example: 5.0
        unit_cost:
          type: number
          format: decimal
          minimum: 0
          default: 0
          description: Costo unitario corrente (può essere 0 per articoli gratuiti)
          example: 8.50
    ItemUpdate:
      type: object
      required:
        - name
        - unit
        - min_stock
        - unit_cost
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: "Filtro abitacolo"
        category:
          type: string
          maxLength: 100
          nullable: true
          example: "Filtri"
        unit:
          type: string
          maxLength: 20
          example: "pz"
        notes:
          type: string
          nullable: true
          example: "Formato universale"
        min_stock:
          type: number
          format: decimal
          minimum: 0
          example: 5.0
        unit_cost:
          type: number
          format: decimal
          minimum: 0
          example: 8.50
    ItemWithStock:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Identificativo univoco
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        name:
          type: string
          example: "Olio motore 5W30"
        category:
          type: string
          nullable: true
          example: "Lubrificanti"
        unit:
          type: string
          example: "lt"
        notes:
          type: string
          nullable: true
          example: "Olio sintetico long life"
        min_stock:
          type: number
          format: decimal
          example: 5.0
        unit_cost:
          type: number
          format: decimal
          description: Costo unitario corrente
          example: 8.50
        stock_quantity:
          type: number
          format: decimal
          description: Giacenza corrente calcolata (somma movimenti)
          example: 17.0
        stock_value:
          type: number
          format: decimal
          description: Valore giacenza (stock_quantity × unit_cost)
          example: 144.50
        is_under_min_stock:
          type: boolean
          description: Flag sotto scorta minima
          example: false
        last_movement_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp ultimo movimento
          example: "2025-11-10T14:30:00Z"
        created_at:
          type: string
          format: date-time
          description: Data creazione articolo
          example: "2025-01-15T10:00:00Z"
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Messaggio errore in italiano
          example: "Impossibile scaricare 5.0 pz. Giacenza disponibile: 3.0 pz"
        error_code:
          type: string
          description: Codice errore machine-readable
          example: "INSUFFICIENT_STOCK"
        context:
          type: object
          nullable: true
          description: Informazioni aggiuntive specifiche per l'errore
          additionalProperties: true
          example:
            requested_quantity: 5.0
            available_stock: 3.0
            unit: "pz"
  parameters:
    ItemId:
      name: id
      in: path
      description: ID univoco dell'articolo
      required: true
      schema:
        type: string
        format: uuid
      example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
  responses:
    BadRequest:
      description: Richiesta non valida (validazione Pydantic fallita)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation_error:
              summary: Errore validazione
              value:
                detail: "Errore di validazione: il campo 'name' è obbligatorio"
                error_code: "VALIDATION_ERROR"
                context:
                  field: "name"
                  constraint: "required"
    NotFound:
      description: Risorsa non trovata
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            item_not_found:
              summary: Articolo non trovato
              value:
                detail: "Articolo con ID 'a1b2c3d4-e5f6-7890-abcd-ef1234567890' non trovato"
                error_code: "ITEM_NOT_FOUND"
            movement_not_found:
              summary: Movimento non trovato
              value:
                detail: "Movimento con ID 'm1a2b3c4-d5e6-7890-abcd-ef1234567890' non trovato"
                error_code: "MOVEMENT_NOT_FOUND"
  securitySchemes:
    # Nessuna autenticazione per MVP (single-user)
    # Placeholder per future implementazioni (JWT, API Key, etc.)
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Autenticazione JWT (non implementata in MVP).
        Per uso futuro con multi-utenza.

# Security: Nessuna autenticazione richiesta per MVP (single-user, accesso diretto)
security: []